refactor: 게임 컴포넌트들  germs 폴더로 이동
