+++ src/components/germs/Player.ts
